moving controls to form directory
